// types/type.d.ts
declare interface Ride {
  id: string;
  origin_address: string;
  destination_address: string;
  created_at: string;
  ride_time: string;
  ride_datetime: string;
  destination_longitude: number;
  destination_latitude: number;
  origin_latitude: number;
  origin_longitude: number;
  available_seats: number;
  driver: {
    first_name: string;
    last_name: string;
    profile_picture: string;
    car_seats: number;
  };
  payment_status: string;
  driver_rating: number;
  status: string;
  driver_id: string;
  recurring: boolean;
  ride_days: string[];
  waypoints: Array<{
    address: string;
    street: string;
    latitude: number;
    longitude: number;
  }>;
  car_type: string;
  priority: number;
  distance: number;
}